Updating due to changes in blackfriday text
This is just a quick implementation of the new TitleBlock required
interface method.

Also, I cloned the blackfriday package and started using the clone so
that changes in blackfriday won't break this package in the future. I'll
have to update things manually, pulling from the canonical blackfriday
and updating blackfridaytext as needed. But that's better than doing a
go get -u and finding out everything is broken.